refactor: move removeImportedToken to confirmation dialog
