+++ frontend/src/lib/components/accounts/IcrcWalletPage.svelte
@@ -21,7 +21,7 @@
   } from "$lib/utils/accounts.utils";
   import { replacePlaceholders } from "$lib/utils/i18n.utils";
   import { IconDots, Island, Spinner } from "@dfinity/gix-components";
-  import type { Principal } from "@dfinity/principal";
+  import { Principal } from "@dfinity/principal";
   import { TokenAmountV2, isNullish, nonNullish } from "@dfinity/utils";
   import type { Writable } from "svelte/store";
   import { ENABLE_IMPORT_TOKEN } from "$lib/stores/feature-flags.store";
@@ -29,8 +29,6 @@
   import WalletMorePopover from "./WalletMorePopover.svelte";
   import { isImportedToken as checkImportedToken } from "$lib/utils/imported-tokens.utils";
   import { importedTokensStore } from "$lib/stores/imported-tokens.store";
-  import { startBusy, stopBusy } from "$lib/stores/busy.store";
-  import { removeImportedTokens } from "$lib/services/imported-tokens.services";
   import ImportTokenRemoveConfirmation from "./ImportTokenRemoveConfirmation.svelte";
   import type { Universe } from "$lib/types/universe";
   import { selectableUniversesStore } from "$lib/derived/selectable-universes.derived";
@@ -168,33 +166,6 @@
   $: universe = $selectableUniversesStore.find(
     ({ canisterId }) => canisterId === ledgerCanisterId?.toText()
   );
-
-  const removeImportedToken = async () => {
-    // Just for type safety. This should never happen.
-    if (isNullish(ledgerCanisterId)) return;
-
-    startBusy({
-      initiator: "import-token-removing",
-      labelKey: "import_token.removing",
-    });
-
-    try {
-      const importedTokens = $importedTokensStore.importedTokens ?? [];
-      const { success } = await removeImportedTokens({
-        tokensToRemove: importedTokens.filter(
-          ({ ledgerCanisterId: id }) =>
-            id.toText() === ledgerCanisterId?.toText()
-        ),
-        importedTokens,
-      });
-
-      if (success) {
-        goto(AppPath.Tokens);
-      }
-    } finally {
-      stopBusy("import-token-removing");
-    }
-  };
 </script>
 
 <TestIdWrapper {testId}>
@@ -274,8 +245,8 @@
   {#if removeImportedTokenConfirmationVisible && nonNullish(universe)}
     <ImportTokenRemoveConfirmation
       {universe}
+      ledgerCanisterId={Principal.fromText(universe.canisterId)}
       on:nnsClose={() => (removeImportedTokenConfirmationVisible = false)}
-      on:nnsConfirm={removeImportedToken}
     />
   {/if}
 </TestIdWrapper>
