+++ frontend/src/lib/components/accounts/ImportTokenRemoveConfirmation.svelte
@@ -5,14 +5,48 @@
   import { nonNullish } from "@dfinity/utils";
   import ConfirmationModal from "$lib/modals/common/ConfirmationModal.svelte";
   import { Tag } from "@dfinity/gix-components";
+  import { startBusy, stopBusy } from "$lib/stores/busy.store";
+  import { importedTokensStore } from "$lib/stores/imported-tokens.store";
+  import { removeImportedTokens } from "$lib/services/imported-tokens.services";
+  import { AppPath } from "$lib/constants/routes.constants";
+  import { goto } from "$app/navigation";
+  import { Principal } from "@dfinity/principal";
+  import { createEventDispatcher } from "svelte";
 
-  export let universe: Universe | undefined;
+  export let universe: Universe | undefined = undefined;
+  export let ledgerCanisterId: Principal;
+
+  const dispatch = createEventDispatcher();
+  const removeImportedToken = async () => {
+    startBusy({
+      initiator: "import-token-removing",
+      labelKey: "import_token.removing",
+    });
+
+    try {
+      const importedTokens = $importedTokensStore.importedTokens ?? [];
+      const { success } = await removeImportedTokens({
+        tokensToRemove: importedTokens.filter(
+          ({ ledgerCanisterId: id }) =>
+            id.toText() === ledgerCanisterId.toText()
+        ),
+        importedTokens,
+      });
+
+      if (success) {
+        dispatch("nnsClose");
+        goto(AppPath.Tokens);
+      }
+    } finally {
+      stopBusy("import-token-removing");
+    }
+  };
 </script>
 
 <ConfirmationModal
   testId="import-token-remove-confirmation-component"
   on:nnsClose
-  on:nnsConfirm
+  on:nnsConfirm={removeImportedToken}
   yesLabel={$i18n.core.remove}
 >
   <div class="content">
